requirements.txt
pyotp==2.9.0
qrcode==8.0
pillow==10.4.0